fix(ci): clear clippy -D warnings debt, feature-gate SPA embed, repair stale tests
Toward the first green CI run on this never-run repo.

Blocker 1 — clippy debt: fix the lints properly (no blanket allows) across core —
unnecessary casts, matches!, is_empty, slice::from_ref, doc list indent — and
collapse three too_many_arguments fns into params/context structs
(MempoolMetricSnapshot, SubmitCtx, events::Submitter).

Blocker 2 — include_dir!(frontend/public) panic: gate the SPA embed behind a
default `embed-spa` feature. `cargo build`/Docker embed as before; the clippy +
test CI jobs build `--no-default-features` so they don't need the frontend
artifact at compile time.

Blocker 3 — stale tests: thread the args the pipelines gained (mempool: entities
+ events; executor/verifier: events/config) and port event assertions from the
old ProviderEvent enum to the struct {kind,payload} model. Restore the verifier's
bundle_failed broadcast (its doc promised it; the enum→struct migration dropped
it). event_watcher_tick/mempool_tick/verifier_tick pass against Postgres.

Still-failing pre-existing drift (separate from the above) tracked next:
single-PP URL drift (/providers/{pk}/ -> /provider/) in several HTTP tests, and
the executor submit-path mock.

+++ crates/api/tests/bundle_submit_http.rs
@@ -4,6 +4,7 @@
 //!   - decodes the MLXDR ops + computes the fee per the provider-platform formula,
 //!   - inserts a PENDING operations_bundles row,
 //!   - returns 201 with bundle_id + status.
+//!
 //! 0-op submissions → 400; missing JWT → 401; non-MLXDR strings → 400.
 
 mod common;
@@ -54,7 +55,7 @@ fn entity_jwt(state: &provider_stack_api::state::AppState) -> String {
 
 fn i128_parts(v: i128) -> Int128Parts {
     Int128Parts {
-        hi: ((v as i128) >> 64) as i64,
+        hi: (v >> 64) as i64,
         lo: ((v as u128) & 0xFFFF_FFFF_FFFF_FFFF) as u64,
     }
 }
